requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
rich>=13.0.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
